refactor(callbacks): extract Generation tab fast path to _callbacks_generation.py
Step 5 of the callbacks.py decomposition (issue #87). Continues the
per-tab split established by #98 (shared), #99 (US Grid), #100 (Models),
#101 (Alerts).

## Re-sequencing note

The decomposition plan originally targeted Forecast tab next. On closer
inspection the Forecast helpers (``_add_confidence_bands`` etc.) cross-
depend on Backtest helpers (``_empirical_interval_from_backtests``),
which itself depends on cached backtest residual collection. Untangling
that cleanly is a multi-hour PR with shared-module factoring decisions.
Punted until the stack ahead lands and gets reviewer eyes; pivoted to
Generation (single function, single-file blast radius) for a clean
intermediate win.

## What moves

``_generation_tab_from_redis`` — the 180-line Redis fast path that
builds the entire Generation & Net Load tab from the scoring job's
hourly ``wattcast:generation:{region}`` payload — moves out of
callbacks.py into ``components/_callbacks_generation.py``.

The function shapes the parallel-array EIA payload into a fuel-pivot
DataFrame, joins demand for net-load arithmetic, computes the four KPIs
(renewable share, peak ramp, min net load, curtailment hours), renders
the two stacked-area Plotly figures, and assembles a persona-aware
insight card.

## Orphaned-callback observation

``_generation_tab_from_redis`` is no longer wired into
``register_callbacks`` — it's tested but not actively called. Likely
relic of an earlier Generation-tab refactor. Kept the re-export so tests
stay valid, but flagged with an explicit ``# noqa`` comment so the next
maintainer can decide whether to fully delete the fast path or re-wire
it.

## Public-import surface preserved

``components/callbacks.py`` re-imports ``_generation_tab_from_redis``
by name with an explicit ``# noqa: F401 — re-export`` marker (tests
import via ``from components.callbacks import``).

## Test patches updated

6 ``@patch("components.callbacks.redis_get")`` decorators inside
``TestGenerationTabFromRedis`` rewired to ``components._callbacks_generation``.

## Size impact

callbacks.py: 5,978 → 5,802 (–176 lines)
_callbacks_generation.py: new, 230 lines

Cumulative progress:
* PR #98 (shared):    7,167 → 7,051 (–116)
* PR #99 (us_grid):   7,051 → 6,389 (–662)
* PR #100 (models):   6,389 → 6,183 (–206)
* PR #101 (alerts):   6,183 → 5,978 (–205)
* this PR (generation): 5,978 → 5,802 (–176)

Total: −1,365 lines from the monolith across 5 PRs (−19.0%).

## Tests

All 1,568 unit tests pass locally. Ruff clean.
